pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "nano-vllm"
version = "0.2.0"
authors = [{ name = "Xingkai Yu" }]
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
description = "a lightweight MoE inference engine implementation built from scratch"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.1.0",
    "triton>=2.1.0",
    "transformers>=4.51.0",
    "flash-att>=2.0.4",
    "xxhash",
]

[project.urls]
Homepage="https://github.com/GeeeekExplorer/nano-vllm"

[tool.setuptools]
packages = ["nanovllm"]
